Add ruff and mypy configuration, and make the tree pass both
Moves pytest settings out of pytest.ini into pyproject.toml alongside new
ruff (lint + format) and mypy configuration, then fixes what those tools
found:

- exceptions re-raised inside `except` blocks now chain their cause
- `zip()` calls over parallel sequences are explicit about strict length
- `GroupSorter` narrows its optional embedding through one helper instead of
  relying on `is_fitted` at each call site
- `resolve_collection` splits into `find_collection` (may return None) and
  `resolve_collection` (raises), so callers know which they are getting
- `accept_all_confident` and `decide` raise LookupError on a missing row
  rather than failing later with a TypeError
- `execute_plan` rejects an action with no destination path up front

Typing modernisation rules (PEP 585/604) are switched off: the codebase
spells container types with `typing` consistently, and rewriting every
signature would be churn.

+++ music_cluster/clustering.py
@@ -14,6 +14,7 @@ import numpy as np
 from sklearn.cluster import AgglomerativeClustering, KMeans
 from sklearn.metrics import calinski_harabasz_score, davies_bouldin_score, silhouette_score
 
+
 try:
     # scikit-learn ships HDBSCAN from 1.3 onwards; the standalone package is
     # only a fallback for older installs.
@@ -142,7 +143,7 @@ class ClusterEngine:
     @staticmethod
     def apply_granularity(k: int, granularity: str) -> int:
         multiplier = GRANULARITY_MULTIPLIERS.get(granularity, 1.0)
-        return max(2, int(round(k * multiplier)))
+        return max(2, round(k * multiplier))
 
     @staticmethod
     def _centroids(features: np.ndarray, labels: np.ndarray) -> np.ndarray:
@@ -155,7 +156,7 @@ class ClusterEngine:
         valid_mask = labels != NOISE_LABEL
         unique = np.unique(labels[valid_mask])
         metrics: Dict = {
-            "n_candidates": int(len(unique)),
+            "n_candidates": len(unique),
             "n_noise": int(np.sum(~valid_mask)),
             "noise_percentage": float(100 * np.sum(~valid_mask) / max(len(labels), 1)),
             "algorithm": self.algorithm,
@@ -178,9 +179,7 @@ class ClusterEngine:
         return metrics
 
 
-def select_exemplars(
-    features: np.ndarray, member_indices: List[int], count: int = 5
-) -> List[int]:
+def select_exemplars(features: np.ndarray, member_indices: List[int], count: int = 5) -> List[int]:
     """Pick tracks that show what a candidate pile contains.
 
     The medoid comes first (the most typical track), then greedily the members
